% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.NBDiagnostics.R
\name{boeplot}
\alias{boeplot}
\title{Baseline/Outcome Event (BOE) plot for NB Models}
\usage{
boeplot(model, diagnostic_stat = c("cookd", "leverage", "anscombe_resid",
  "dfbeta"), breaks = NULL, n.label = 3, id_lable_size = 5,
  show_labels_id = NULL)
}
\arguments{
\item{model}{an object of class \code{NBDiagnostics}}

\item{diagnostic_stat}{the diagnostic statistic that should be presented in the BOE plot}

\item{n.label}{the number of ID labels shown in the plot}

\item{id_lable_size}{the size of the ID labels}

\item{show_labels_id}{the specified IDs to show in the labels}
}
\description{
The Baseline/Outcome Event (BOE) plot for a fitted NB model
}
\examples{
## simulate a dataset with outcome and baseline count
obs <- 200
s <- rgamma(obs, shape = 1/3, scale = 3)
y0 <- rpois(obs, 25 * s)
x <- rep(c(1, 0), each=obs/2)
y1 <- rpois(obs, 25 * s * exp(-0.7 * x))
x <- factor(x, levels = c(1, 0),
            labels = c('Intervention', 'Control'))
dat <- data.frame(id=1:obs, y0=y0, y1=y1, group=x)

## fit an NB model using nbdiagnostics() and specify
## the varaibles names for the plots
mod_nb <- nbdiagnostics(
  y1 ~ group, data=dat, id_varname = "id",
  outcome_varname="y1", baseline_varname="y0",
  group_varname = "group")

## produce the baseline/outcome event plots
boeplot(mod_nb, diagnostic_stat = "cookd")
boeplot(mod_nb, diagnostic_stat = "leverage")
boeplot(mod_nb, diagnostic_stat = "anscombe_resid")
boeplot(mod_nb, diagnostic_stat = "dfbeta")
}
