% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.NBDiagnostics.R
\name{caprob_nb}
\alias{caprob_nb}
\title{Covariate adjusted probability plot for NB models}
\usage{
caprob_nb(len, model)
}
\arguments{
\item{len}{the covariate adjusted probabilities within 0 to \code{len}}

\item{model}{an object of class \code{NBDiagnostics}}
}
\description{
The covariate adjusted probability plot (Holling et al, 2016) of the NB model.
}
\examples{
## simulate a dataset with outcome and baseline count
obs <- 200
s <- rgamma(obs, shape = 1/3, scale = 3)
y0 <- rpois(obs, 25 * s)
x <- rep(c(1, 0), each=obs/2)
y1 <- rpois(obs, 25 * s * exp(-0.7 * x))
x <- factor(x, levels = c(1, 0),
            labels = c('Intervention', 'Control'))
dat <- data.frame(id=1:obs, y0=y0, y1=y1, group=x)

## fit an NB model using nbdiagnostics() and specify
## the varaibles names for the plots
mod_nb <- nbdiagnostics(
  y1 ~ group, data=dat, id_varname = "id",
  outcome_varname="y1", baseline_varname="y0",
  group_varname = "group")

## covariate adjusted probability plot: NB (Holling et al, 2016)
caprob_nb(20, mod_nb)
}
\references{
Holling, H., Böhning, W., Böhning, D., Formann, A.K., 2016. The covariate-adjusted frequency plot. Stat. Methods Med. Res. 25, 902–916.
}
