% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBDiagnostics.R
\name{nbdiagnostics}
\alias{nbdiagnostics}
\title{Diagnostic functionality for NB models}
\usage{
nbdiagnostics(formula, data, ..., id_varname, outcome_varname,
  baseline_varname, group_varname, addition = 0.5)
}
\arguments{
\item{formula, data, ...}{arguments for the \code{glm.nb()} function}

\item{id_varname}{variable name for subject ID}

\item{outcome_varname}{variable name for the outcome count/rate}

\item{baseline_varname}{variable name for the baseline count/rate}

\item{group_varname}{variable name for group}

\item{addition}{addition for log-transformation}
}
\value{
\code{nbdiagnostics()} returns an object of class \code{NBDiagnostics}
inheriting from \code{negbin}, \code{glm}, and \code{lm}.
}
\description{
Diagnostic functionality for NB models
}
\examples{
## simulate a dataset with outcome and baseline count
obs <- 200
s <- rgamma(obs, shape = 1/3, scale = 3)
y0 <- rpois(obs, 25 * s)
x <- rep(c(1, 0), each=obs/2)
y1 <- rpois(obs, 25 * s * exp(-0.7 * x))
x <- factor(x, levels = c(1, 0),
            labels = c('Intervention', 'Control'))
dat <- data.frame(id=1:obs, y0=y0, y1=y1, group=x)

## fit an NB model using nbdiagnostics() and specify
## the varaibles names for the plots
mod_nb <- nbdiagnostics(
  y1 ~ group, data=dat, id_varname = "id",
  outcome_varname="y1", baseline_varname="y0",
  group_varname = "group")

## produce the baseline/outcome event plots
boeplot(mod_nb, diagnostic_stat = "cookd")
boeplot(mod_nb, diagnostic_stat = "leverage")
boeplot(mod_nb, diagnostic_stat = "anscombe_resid")
boeplot(mod_nb, diagnostic_stat = "dfbeta")

## covariate adjusted probability plots (Holling et al, 2016)
caprob_nb(20, mod_nb)
caprob_nb_poi(20, mod_nb, data=dat)
}
